
\name{chromEvol-internal}
\alias{read.ancProb.tab}
\alias{read.expect.tab}
\alias{read.mlAnc.tree}
\alias{read.modSum}
\alias{read.param}
\alias{read.PP.tree}
\alias{read.results}
\alias{specify.model}
\alias{write.param}
\alias{get.PP}
\alias{match.table2tip}
\title{Internal ChromEvol Functions}
\description{
 Internal ChromEvol functions
}
\Usage{ only for internal use
}
\details{
These are not to be called by the user.
}
\keyword{ internal }
